Refactored Code Created Adaptor Factory for achieving encapsulation

+++ src/main/java/censusanalyser/USCensusCSV.java
@@ -16,5 +16,5 @@ public class USCensusCSV {
     public double totalArea;
 
     @CsvBindByName(column = "Population Density",required = true)
-    public int populationDensity;
+    public double populationDensity;
 }
